requirements.txt
# requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
motor>=3.3.1
pydantic>=2.4.2
pydantic-settings>=2.0.3 # For easy .env loading
python-dotenv>=1.0.0    # Dependency for pydantic-settings
bson>=0.5.10